requirements.txt
Django
djangorestframework==3.12.4
pillow
django-filter
pytz
gunicorn
psycopg2-binary
webcolors
flake8